refactor: remove redundant cast to int

+++ app/Actions/WaitlistSignups/CountWaitlistSignupsAction.php
@@ -12,7 +12,7 @@ final class CountWaitlistSignupsAction
     public function __invoke(): int
     {
         /** @phpstan-ignore cast.int */
-        return (int) Cache::remember(
+        return Cache::remember(
             'waitlistSignupsCount',
             now()->addWeek(),
             static fn () => WaitlistSignup::query()->count()
